Enable file upload by passing file content data
When submitting a form, currently Goutte Client can upload a file only using an existing file path, so I added the use of a 'content' element for file info array in the request. In this way the Client can upload a file by passing file content as a string, a stream resource or an object.